pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cypulse"
version = "0.1.0"
description = "開源 EASM 資安曝險評級平台"
readme = {text = "CyPulse — 開源 EASM 資安曝險評級平台", content-type = "text/plain"}
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31",
    "structlog>=23.1",
    "click>=8.1",
    "jinja2>=3.1",
    "weasyprint>=60.0",
    "checkdmarc>=5.0",
    "dnstwist>=20230918",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "flake8>=6.0",
    "black>=23.0",
    "mypy>=1.0",
]

[project.scripts]
cypulse = "cypulse.cli:main"

[tool.setuptools.packages.find]
include = ["cypulse*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
